+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 -   [setup]: Unit tests
 -   [added]: Scripts to mount evirontments
 -   [added]: Input base component.
+-   [fixed]: Dark and light mode for input